Include new field responses.
These are the ones, which doesn't have fieldResponseId. Normally, when a record is created, initial field response are automatically created, which is then later used for updating/saving the form response. But, sometimes, new field might be added (after app update) after the record creation. In such scenario, we can use check newFieldResponses and handle it accordingly in the backend i.e. create new response using it.