Merged PR 11577: Stage Advancement email
This PR enables email notifications upon stage advancement. For example, when a student submits a plan or literature review, the instructor receives a submission notification. Similarly, when an instructor requests changes or approves a plan, the student receives an email.

Additionally, it updates the tests, including a fixture for **`StageService`**, and organises fixtures into a separate `Fixtures` folder.