Merged PR 11766: Create missing field response upon form submission
This PR fixes an issue where missing field responses are not generated.

For e.g.. when a new fields are seeded into the section after a plan has been created, the field responses for these new fields are not available. Saving the section form should generate these missing field responses, but this is not happening currently. This PR addresses and resolves that issue.

Other changes include:
- Updating the **`CreateFieldResponse`** method in `SectionFormService` to prevent duplicate field response creation.