Merged PR 12000: Redirect to section form when there is only one section
The PR updates the existing workflow where a user is redirected to the section form directly when only one section is available. For e.g.. currently, when a user tries to view a literature review page, they are navigated to the overview page, which only has one section. This update improves the UX.

Other updates include:-
- Add new functions **`buildOverviewPath`**, **`buildSectionFormPath`** and **`buildProjectPath`** that returns relevant route paths accordingly. This helps in maintaining consistency.
- Breadcrumb fixes (hide link to overview in section form if no more than one section)
- Updates the **`DraggableListField`** component to add `isDisabled` prop

Related work items: #178326, #178327, #178328